app.py
import streamlit as st
import pandas as pd
import numpy as np
from dotenv import load_dotenv
load_dotenv()

# Import other necessary libraries from your notebook
# Copy the imports from the notebook script

st.title("GenAI-Driven Market Comparison (using Zhipu AI GLM)")

st.markdown("""
This is a Streamlit app based on the GenAI-Driven Market Comparison notebook.
It demonstrates how Generative AI (using Zhipu AI GLM model) can streamline market comparisons in the insurance industry.
""")

# Add interactive elements
# For example, file uploader for PDFs
uploaded_file = st.file_uploader("Upload a PDF report", type="pdf")

if uploaded_file is not None:
    # Process the uploaded file
    # Add your processing logic here
    st.write("File uploaded successfully!")
    # Then run your analysis and display results

# Alternatively, run the analysis on predefined data
if st.button("Run Analysis"):
    # Import and run parts of your notebook code
    # For example:
    # from your_script import some_function
    # results = some_function()
    # st.write(results)

    st.write("Analysis results will be displayed here.")

# Display static results or tables from the notebook
# You can hardcode or load pre-computed results

st.header("Key Insights")
st.markdown("""
- RAG framework effectively extracts relevant sections.
- Structured Outputs ensure consistent formatting.
- Actuarial expertise is essential for validation.
""")

# Add more sections as needed